"Calculator with more tests"

+++ L2-Zad.3-Kalkulator/src/main/java/project/Calculator.java
@@ -1,16 +1,16 @@
 package project;
 
 public class Calculator {
-    public static int sum(int a, int b){
-        return a + b;
+    public static long sum(int a, int b){
+        return (long)a + (long)b;
     }
-    public static int subtract(int a, int b){
-        return a - b;
+    public static long subtract(int a, int b){
+        return (long)a - (long)b;
     }
-    public static int multiply(int a, int b){
-        return a * b;
+    public static long multiply(int a, int b){
+        return (long)a * (long)b;
     }
-    public static int divide(int a, int b) {
-        return a / b;
+    public static double divide(int a, int b) {
+        return (double)a / (double)b;
     }
 }
